Add files via upload
I am providing you with the script for the first assignment of the Spring 2023 semester, along with the accompanying dataset used for this task. This assignment, like others, has been developed and evaluated under the guidance of our professors. Feel free to review the script and explore the dataset as needed.